modules/LiveDataCollector/Ctrader/requirements.txt
ctrader-open-api
ctrader-fix
simplefix==1.0.17
redis
psycopg2-binary